% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let2num.R
\name{let2num}
\alias{let2num}
\title{De letras a numeros}
\usage{
let2num(x, lets = LETTERS)
}
\arguments{
\item{x}{es un vector que detalla los nombres a utilizar.}

\item{lets}{es un vector que detalla los nombres a utilizar.}
}
\description{
Esta es una funcion de ejemplo para multiplicar un valor o vector por dos.
}
\details{
Esta funcion es parte del tutorial para hacer paquetes del curso de Metodos Multivariados 2017
en el ITAM. La funcion recibe un elemento que puede ser un numero o un vector numerico y lo devuelve
multiplicado por dos. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
num2let(x = 1)
num2let(x = c(3, 10, 256, 1000, 2))
}
