% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namel2.R
\name{namel2}
\alias{namel2}
\title{Crear una lista con nombres predeterminados}
\usage{
namel2(names, vec = NA)
}
\arguments{
\item{names}{es un vector que detalla los nombres a utilizar.}
}
\description{
Esta es una funcion de ejemplo para multiplicar un valor o vector por dos.
}
\details{
Esta funcion es parte del tutorial para hacer paquetes del curso de Metodos Multivariados 2017
en el ITAM. La funcion recibe un elemento  que puede ser un numero o un vector numerico y lo devuelve
multiplicado por dos. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
namel2(names = colnames(mtcars))
namel2(names = colnames(mtcars), vec = colnames(mtcars))
}
