% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2let.R
\name{num2let}
\alias{num2let}
\title{De numeros a letras}
\usage{
num2let(n, lets = LETTERS)
}
\arguments{
\item{n}{es un vector numerico con valores mayores o iguales a 1.}

\item{lets}{es un vector que detalla las letras del abecedario en mayusculas.}
}
\description{
La función \code{namel2()} realiza la conversion de numero a letra o combinacion de letras
correspondientes al abecedario latino.
}
\details{
El vector \code{n} solo soporta valores mayores o igual a 1 como parte del vector numerico.
Ante valores NULL, NA, valores numericos 0 o negativos, la funcion dara un mensaje de error.
El vector \code{lets} no soporta valores \code{NA} o NULL entre sus elementos.
}
\examples{
num2let(1)
num2let(c(3, 10, 256, 1000, 2))
}
