% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Redondeo Humano}
\usage{
round2(x, decimals = 4)
}
\arguments{
\item{x}{es un vector numerico, una matriz numerica o un data.frame numerico.}

\item{decimals}{es un vector numerico de un unico elemento que indica la cantidad de decimales a utilizar.}
}
\description{
La función \code{round2()} es una funcion que realiza la tarea de redondear numeros
de la manera en que estamos acostumbrados a realizarlo. Es una opcion diferente al redondeo que ofrece
la funcion round() del paquete base de R.
}
\details{
Realiza el redondeo de acuerdo a la cantidad de dígitos y los valores.
Si \code{x > 0} y el digito en la posicion decimal a redondear es mayor o igual a 5, redondea hacia
el valor positivo siguiente.
Si \code{x < 0} y el digito en la posicion decimal a redondear es mayor o igual a 5, redondea hacia
Si \code{x < 0}el valor negativo anterior
Si Si \code{x == 0}, se redondea a 0.
}
\examples{
round(0.5)
round2(0.5)
}
