% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TodoJunto.R
\name{Rscience.languages}
\alias{Rscience.languages}
\title{Todo junto}
\usage{
Rscience.languages(name = NULL)
}
\arguments{
\item{names}{es un vector que detalla los nombres a asignar a la lista.}

\item{vec}{es un vector que detalla el contenido minimo que tendra cada lista.
Puede ser un vector de un solo elemento o un vector con un elemento para cada lista.
El valor por defecto es NA.}
}
\description{
Un monton de cosas todas juntas.
}
\details{
Internamente, genera una lista con los elementos del vector \code{vec}, y le asigna a la lista
generada los nombres del vector \code{names}. El parametro \code{vec} por defecto es NA.
Si se asigna un vector a \code{vec} este vector debe contener un elemento unico o tener
la misma cantidad de elementos que \code{names}; en cualquier otor caso la funcion devuelve un aviso de error.
}
\examples{
namel2(names = colnames(mtcars))
namel2(names = colnames(mtcars), vec = colnames(mtcars))
}
