% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2let.R
\name{num2let}
\alias{num2let}
\title{De numeros a letras}
\usage{
num2let(n, lets = LETTERS)
}
\arguments{
\item{n}{es un vector que detalla los nombres a utilizar.}
}
\description{
Esta es una funcion de ejemplo para multiplicar un valor o vectorpor dos.
}
\details{
Esta funcion es parte del tutorial para hacer paquetes del curso de Metodos Multivariados 2017
en el ITAM. La funcion recibe un elemento que puede ser un numero o un vector numerico y lo devuelve
multiplicado por dos. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
num2let(1)
num2let(c(3, 10, 256, 1000, 2))
}
