% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Redondea como lo hacemos en la primaria}
\usage{
round2(x, n)
}
\arguments{
\item{x}{es un vector que detalla los nombres a utilizar.}
}
\description{
Esta es una funcion de ejemplo para multiplicar un valor o vector por dos.
}
\details{
Esta funcion es parte del tutorial para hacer paquetes del curso de Metodos Multivariados 2017
en el ITAM. La funcion recibe un elemento que puede ser un numero o un vector numerico y lo devuelve
multiplicado por dos. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
round(0.5)
round2(0.5)
}
